+++ modules/farm_rothamsted_researcher/src/Entity/RothamstedResearcher.php
@@ -285,26 +285,6 @@ class RothamstedResearcher extends RevisionableContentEntityBase implements Roth
         'label' => 'inline',
       ]);
 
-    $fields['notification_enabled'] = BaseFieldDefinition::create('boolean')
-      ->setLabel(t('Notifications'))
-      ->setDescription(t('Enable email notifications for the researcher.'))
-      ->setDefaultValue(TRUE)
-      ->setRevisionable(TRUE)
-      ->setRequired(TRUE)
-      ->setSettings([
-        'on_label' => t('Enabled'),
-        'off_label' => t('Disabled'),
-      ])
-      ->setDisplayConfigurable('form', TRUE)
-      ->setDisplayOptions('form', [
-        'type' => 'options_buttons',
-      ])
-      ->setDisplayConfigurable('view', TRUE)
-      ->setDisplayOptions('view', [
-        'type' => 'boolean',
-        'label' => 'inline',
-      ]);
-
     $fields['job_title'] = BaseFieldDefinition::create('string')
       ->setLabel(t('Job title'))
       ->setDescription(t('The job title of the researcher.'))
@@ -408,6 +388,16 @@ class RothamstedResearcher extends RevisionableContentEntityBase implements Roth
       ]);
 
     return $fields;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getNotificationEmail(): ?string {
+    if ($this->get('farm_user')->isEmpty()) {
+      return NULL;
+    }
+    return $this->get('farm_user')->entity->get('rothamsted_notification_email')->value ? $this->get('farm_user')->entity->get('mail')->value : NULL;
   }
 
 }